+++ view/frontend/web/js/view/payment/next.js
@@ -49,17 +49,6 @@ define([
         return paymentValidatorList[paymentComponent.getCode()](paymentComponent);
     }
 
-    function afterBillingAddressSet() {
-        var selectedPaymentMethodComponent = getSelectedPaymentMethodComponent();
-        if (selectedPaymentMethodComponent) {
-            if (validatePayment(selectedPaymentMethodComponent)) {
-                setPaymentInformation(globalMessageList, selectedPaymentMethodComponent.getData()).done(function() {
-                    stepNavigator.next();
-                });
-            }
-        }
-    }
-
     return Component.extend({
         initialize: function () {
             var self = this;
@@ -67,25 +56,34 @@ define([
             quote.paymentMethod.subscribe(function () {
                 self.errorValidationMessage(false);
             });
+            quote.billingAddress.subscribe(function () {
+                self.errorValidationMessage(false);
+            });
         },
         errorValidationMessage: ko.observable(false),
         setBillingInformation: function() {
-            // @todo Save payment method!
+            // @todo Save payment method?
             if (this.validateBillingInformation()) {
-                if (getBillingAddressComponent().isAddressDetailsVisible()) {
-                    afterBillingAddressSet();
-                } else {
-                    getBillingAddressComponent().updateAddress().done(function() {
-                        afterBillingAddressSet();
-                    });
+                var selectedPaymentMethodComponent = getSelectedPaymentMethodComponent();
+                if (selectedPaymentMethodComponent) {
+                    if (validatePayment(selectedPaymentMethodComponent)) {
+                        setPaymentInformation(globalMessageList, selectedPaymentMethodComponent.getData())
+                            .done(function() {
+                                stepNavigator.next();
+                            });
+                    }
                 }
             }
         },
         validateBillingInformation: function() {
             var loginFormSelector = 'form[data-role=email-with-possible-login]',
-                billingAddressFormSelector = '.billing-address-form form',
-                billingAddressValidationResult = true,
                 emailValidationResult = customer.isLoggedIn();
+            if (!getBillingAddressComponent().isAddressDetailsVisible()) {
+                this.errorValidationMessage(
+                    $t('First you need to save billing address.')
+                );
+                return false;
+            }
             if (!quote.paymentMethod()) {
                 this.errorValidationMessage(
                     $t('The payment method is missing. Select the payment method and try again.')
@@ -95,17 +93,9 @@ define([
             if (!customer.isLoggedIn()) {
                 $(loginFormSelector).validation();
                 emailValidationResult = Boolean($(loginFormSelector + ' input[name=username]').valid());
-                if (!getBillingAddressComponent().isAddressDetailsVisible()) {
-                    $(billingAddressFormSelector).validation();
-                    billingAddressValidationResult = getBillingAddressComponent().validateAddress();
-                }
             }
             if (!emailValidationResult) {
                 $(loginFormSelector + ' input[name=username]').focus();
-                return false;
-            }
-            if (!billingAddressValidationResult) {
-                $(billingAddressFormSelector + ' .field._error:first input').focus();
                 return false;
             }
             return true;